+++ Scripts/Buttons/Editor/NavContainerPropDrawer.cs
@@ -1,19 +1,24 @@
-#if UNITY_EDITOR
 using UnityEditor;
 using UnityEditor.UIElements;
 using UnityEngine.UIElements;
+using static DevPeixoto.UI.MenuManager.UGUI.ButtonMenuNav;
 
 namespace DevPeixoto.UI.MenuManager.UGUI
 {
-    [CustomEditor(typeof(ButtonMenuNav))]
-    public class ButtonMenuNavEditor : Editor
+    [CustomPropertyDrawer(typeof(NavContainer))]
+    public class NavContainerPropDrawer : PropertyDrawer
     {
-        public override VisualElement CreateInspectorGUI()
+        public override VisualElement CreatePropertyGUI(SerializedProperty prop)
         {
             var root = new VisualElement();
 
+            ObjectField selfField = new ObjectField();
+            selfField.value = prop.FindPropertyRelative("selfRef").objectReferenceValue;
+            selfField.SetEnabled(false);
+            root.Add(selfField);
+
             ObjectField ownerField = new ObjectField("Respective Menu Owner");
-            var ownerProp = serializedObject.FindProperty("owner");
+            var ownerProp = prop.FindPropertyRelative("owner");
             ownerField.value = ownerProp.objectReferenceValue;
             ownerField.SetEnabled(false);
             ownerField.style.marginTop = 5;
@@ -21,7 +26,7 @@ namespace DevPeixoto.UI.MenuManager.UGUI
             root.Add(ownerField);
 
             DropdownField targetMenuDropDownField = new DropdownField("Target Menu");
-            var targetMenuProp = serializedObject.FindProperty("targetMenu");
+            var targetMenuProp = prop.FindPropertyRelative("targetMenu");
             targetMenuDropDownField.BindProperty(targetMenuProp);
             targetMenuDropDownField.choices = MenuOptionsSingleton.Instance.GetOptions(ownerProp.objectReferenceValue as MenusManager);
             targetMenuDropDownField.choices.Insert(0, "None");
@@ -29,8 +34,12 @@ namespace DevPeixoto.UI.MenuManager.UGUI
             targetMenuDropDownField.style.marginBottom = 5;
             root.Add(targetMenuDropDownField);
 
+            Toggle openOverlayToggle = new Toggle("Open as Overlay");
+            var overlayProp = prop.FindPropertyRelative("Overlay");
+            openOverlayToggle.BindProperty(overlayProp);
+            root.Add(openOverlayToggle);
+
             return root;
         }
     }
 }
-#endif